DCMAW-9664: Rename state configuaration class for async token
